Factorized certificate creation (RSA or self-signed)
