+++ generators/python-bots/index.js
@@ -1,8 +1,5 @@
 const Generator = require('yeoman-generator');
-const colors = require('colors');
-const certificateCreator = require('../lib/p12-certificate-creator');
-const RSAcertificateCreator = require('../lib/certificate-creator');
-var mkdirp = require('mkdirp');
+const CertificateCreator = require('../lib/certificate-creator');
 
 module.exports = class extends Generator {
     prompting() {
@@ -100,18 +97,7 @@ module.exports = class extends Generator {
                 );
             }
             /* Install certificate */
-            /*TODO add logging */
-            if (answers.encryption === 'Self Signed Certificate') {
-                let log_text_cert = ('* Generating certificate for BOT ' + answers.botusername + '...').bold;
-                console.log(log_text_cert.bgRed.white);
-                mkdirp.sync('certificates');
-                certificateCreator.create(answers.botusername, 'certificates');
-            } else if (answers.encryption === 'RSA - Generate New Keys') {
-                let log_text_cert = ('* Generating RSA public/private keys for BOT ' + answers.botusername + '...').bold;
-                console.log(log_text_cert.bgRed.white);
-                mkdirp.sync('rsa');
-                RSAcertificateCreator.createRSA(answers.botusername, 'rsa');
-            }
+            CertificateCreator.create(this.options.initPrompts.encryption, answers.botusername, answers.botemail);
 
             let log_text_completion = ('* BOT generated successfully!!').bold;
             console.log(log_text_completion.bgGreen.white);
